requirements.txt
Django==1.7.4
BeautifulSoup==3.2.1
Markdown==2.4.1
Pillow==2.5.3
PyICU==1.8
celery==3.1.14
django-analytical==0.19.0
django-axes==1.3.8
django-celery==3.1.10
django-cors-headers==0.13
-e git+https://github.com/toastdriven/django-haystack.git@01c234e4d0013e17b31d3a1d593e05e5f340d462#egg=django_haystack-dev
django-ipware
django-jsonfield==0.9.13
django-modeltranslation==0.7.3
django-mptt==0.6.1
django-redis==3.7.1
django-taggit==0.12.1
djangorestframework==2.3.14
djmail==0.8
easy-thumbnails==2.1
kombu==3.0.22
python-dateutil
python-openid==2.2.5
python-social-auth==0.1.26
redis==2.10.3
git+https://github.com/notanumber/xapian-haystack.git
git+https://github.com/pennersr/django-cooki18n.git
python-slugify==0.1.0
django-postman==3.2.0
django-simple-captcha==0.4.4
django-activity-stream==0.5.1
